pyproject.toml
[tool.poetry]
name = "fastblog"
version = "1.7.0"
description = "A RESTful blog system built with Python and FastAPI ⚡"
authors = ["MahdiHaghverdi <mahdihaghverdliewpl@gmail.com>"]
license = 'GPLv3'
readme = "README.md"
repository = 'https://github.com/mahdihaghverdi/FastBlog'

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {version = "~0.103.1", extras = ["all"]}
alembic = "~1.12.0"
pyjwt = "~2.8.0"
cryptography = "~41.0.3"
sqlalchemy = {version = "~2.0.20", extras = ["asyncio"]}
asyncpg = "~0.28.0"
python-jose = {extras = ["cryptography"], version = "~3.3.0"}
passlib = {extras = ["bcrypt"], version = "~1.7.4"}

[tool.poetry.group.dev.dependencies]
pre-commit = "~3.3.3"


[tool.poetry.group.test.dependencies]
pytest = "~7.4.1"
aiosqlite = "^0.19.0"
pytest-cov = "^4.1.0"

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/mahdihaghverdi/FastBlog/issues"

[tool.coverage.report]
exclude_also = [
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
    "pass",
    ]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
